Streamlines test classes to make their purpose more apparent.
