--- embabel-agent-test-support/embabel-agent-test/src/test/java/com/embabel/agent/test/integration/ExampleAgentApplication.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2024-2026 Embabel Pty Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.embabel.agent.test.integration;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ExampleAgentApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ExampleAgentApplication.class, args);
-    }
-}
