--- embabel-agent-test-support/embabel-agent-test/src/test/java/com/embabel/agent/test/integration/WriteAndReviewAgentIntegrationTest.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2024-2026 Embabel Pty Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.embabel.agent.test.integration;
-
-import com.embabel.agent.api.invocation.AgentInvocation;
-import com.embabel.agent.domain.io.UserInput;
-import com.embabel.agent.test.integration.WriteAndReviewAgent.ReviewedStory;
-import org.junit.jupiter.api.Test;
-
-import static com.embabel.agent.test.integration.WriteAndReviewAgent.Story;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-/**
- * Use framework superclass to test the complete workflow of writing and reviewing a story.
- * This will run under Spring Boot against an AgentPlatform instance
- * that has loaded all our agents.
- */
-class WriteAndReviewAgentIntegrationTest extends EmbabelMockitoIntegrationTest {
-
-    @Test
-    void shouldExecuteCompleteWorkflow() {
-        var input = new UserInput("Write about artificial intelligence");
-
-        var story = new Story("AI will transform our world...");
-        var reviewedStory = new ReviewedStory(story, "Excellent exploration of AI themes.", Personas.REVIEWER);
-
-        whenCreateObject(s -> s.contains("Craft a short story"), Story.class)
-                .thenReturn(story);
-
-        // The second call uses generateText
-        whenGenerateText(s -> s.contains("You will be given a short story to review"))
-                .thenReturn(reviewedStory.review());
-
-        var invocation = AgentInvocation.create(agentPlatform, ReviewedStory.class);
-        var reviewedStoryResult = invocation.invoke(input);
-
-        assertNotNull(reviewedStoryResult);
-        assertTrue(reviewedStoryResult.story().text().contains(story.text()),
-                "Expected story content to be present: " + reviewedStoryResult.story().text());
-        assertEquals(reviewedStory, reviewedStoryResult,
-                "Expected review to match: " + reviewedStoryResult);
-
-        verifyCreateObjectMatching(prompt -> prompt.contains("Craft a short story"), Story.class,
-                llm -> llm.getLlm().getTemperature() != null && llm.getLlm().getTemperature() == 0.9 && llm.getToolGroups().isEmpty());
-        verifyGenerateTextMatching(prompt -> prompt.contains("You will be given a short story to review"));
-        verifyNoMoreInteractions();
-    }
-}
